+++ src/main/java/net/silentchaos512/gear/api/item/ICoreItem.java
@@ -28,8 +28,6 @@ public interface ICoreItem extends IItemProvider, IStatItem {
         ItemStack result = new ItemStack(this);
         GearData.writeConstructionParts(result, parts);
         GearData.recalculateStats(result, null);
-        // Allow upgrade parts to add additional data
-        parts.forEach(p -> p.onAddToGear(result));
         // Allow traits to make any needed changes (must be done after a recalculate)
         TraitHelper.activateTraits(result, 0, (trait, level, nothing) -> {
             trait.onGearCrafted(new TraitActionContext(null, level, result));