+++ src/main/java/net/silentchaos512/gear/crafting/recipe/GearPartSwapRecipe.java
@@ -6,19 +6,23 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.ICraftingRecipe;
 import net.minecraft.item.crafting.IRecipeSerializer;
 import net.minecraft.network.PacketBuffer;
+import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
 import net.minecraftforge.registries.ForgeRegistryEntry;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.item.ICoreItem;
+import net.silentchaos512.gear.api.parts.IGearPart;
 import net.silentchaos512.gear.api.parts.PartDataList;
 import net.silentchaos512.gear.api.parts.PartType;
 import net.silentchaos512.gear.parts.PartData;
+import net.silentchaos512.gear.parts.PartManager;
 import net.silentchaos512.gear.util.GearData;
 import net.silentchaos512.lib.collection.StackList;
 
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 
 public class GearPartSwapRecipe implements ICraftingRecipe {
     public static final ResourceLocation NAME = new ResourceLocation(SilentGear.MOD_ID, "swap_gear_part");
@@ -67,7 +71,6 @@ public class GearPartSwapRecipe implements ICraftingRecipe {
             if (part == null) return ItemStack.EMPTY;
 
             // Remove old part of type, replace
-            // TODO: Can we return the old part? Probably not reliably with vanilla crafting...
             PartType type = part.getType();
             parts.removeIf(p -> p.getType() == type);
             parts.add(part);
@@ -76,6 +79,36 @@ public class GearPartSwapRecipe implements ICraftingRecipe {
         GearData.writeConstructionParts(result, parts);
         GearData.recalculateStats(result, null);
         return result;
+    }
+
+    @Override
+    public NonNullList<ItemStack> getRemainingItems(CraftingInventory inv) {
+        NonNullList<ItemStack> list = NonNullList.withSize(inv.getSizeInventory(), ItemStack.EMPTY);
+        StackList stackList = StackList.from(inv);
+        ItemStack gear = stackList.uniqueMatch(s -> s.getItem() instanceof ICoreItem);
+        PartDataList oldParts = GearData.getConstructionParts(gear);
+
+        for (int i = 0; i < list.size(); ++i) {
+            ItemStack stack = inv.getStackInSlot(i);
+
+            if (stack.getItem() instanceof ICoreItem) {
+                list.set(i, ItemStack.EMPTY);
+            } else {
+                IGearPart part = PartManager.from(stack);
+                if (part != null) {
+                    List<PartData> partsOfType = oldParts.getPartsOfType(part.getType());
+                    if (!partsOfType.isEmpty()) {
+                        PartData partData = partsOfType.get(0);
+                        partData.onRemoveFromGear(gear);
+                        list.set(i, partData.getCraftingItem());
+                    } else {
+                        list.set(i, ItemStack.EMPTY);
+                    }
+                }
+            }
+        }
+
+        return list;
     }
 
     @Override