+++ src/main/java/net/silentchaos512/gear/util/GearGenerator.java
@@ -118,7 +118,6 @@ public final class GearGenerator {
         }
 
         ItemStack result = stack.copy();
-        parts.forEach(p -> p.onAddToGear(result));
         GearData.writeConstructionParts(result, parts);
 
         // Apply some random upgrades?