+++ CHANGELOG.md
@@ -11,8 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (API) PartTypes can now specify a fallback part, deprecates `PartManager.tryGetFallback`
 - (API) Gear class maps in ModItems now use ResourceLocations for keys, to allow other mods to add gear types
 ### Fixed
+- Swapping gear parts will now attempt to return the replaced parts in the crafting grid
 - Part names being underlined after viewing them in the parts menu
 - Gear randomizer not selecting parts in some cases
+- Random gear command not working with command blocks
 - Salvager attaching grade NBT to returned items
 
 ## [1.5.8] - 2020-04-15